requirements.txt
discord.py==2.6.4
python-dotenv==1.2.1
watchfiles==1.1.1
